# docs/openapi.yaml
openapi: 3.0.3
info:
  title: PCCraft Marketplace API
  version: 1.0.0
  description: REST API for the PCCraft Marketplace platform -- multi-vendor PC components
    e-commerce with recommendations and a data-driven compatibility engine.
paths:
...
  /api/v1/admin/reviews/:
    get:
      operationId: admin_reviews_list
      description: '``/api/v1/admin/reviews/`` family.'
      parameters:
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      tags:
      - admin
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAdminReviewListList'
          description: ''
...
  /api/v1/admin/reviews/{id}/hide/:
    post:
      operationId: admin_reviews_hide_create
      description: '``/api/v1/admin/reviews/`` family.'
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - admin
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminReviewList'
          description: ''
  /api/v1/admin/reviews/{id}/moderate/:
    patch:
      operationId: admin_reviews_moderate_partial_update
      description: |-
        Spec: ``PATCH /admin/reviews/{id}/moderate/`` body ``{is_hidden}``.

        Single endpoint that toggles ``is_hidden`` to the requested
        value.  Re-aggregates the product average after every change.
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - admin
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminReviewList'
          description: ''
  /api/v1/admin/reviews/{id}/reply/:
    delete:
      operationId: admin_reviews_reply_destroy
      description: '``/api/v1/admin/reviews/`` family.'
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - admin
      security:
      - jwtAuth: []
      responses:
        '204':
          description: No response body
  /api/v1/admin/reviews/{id}/restore/:
    post:
      operationId: admin_reviews_restore_create
      description: '``/api/v1/admin/reviews/`` family.'
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - admin
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminReviewList'
          description: ''
...
  /api/v1/products/{slug}/reviews/:
    get:
      operationId: products_reviews_list
      description: |-
        ``/api/v1/products/{slug}/reviews/`` family.

        GenericViewSet (not ModelViewSet) because the resource is nested
        under product -- we don't want a top-level list view here.
      parameters:
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      - in: path
        name: slug
        schema:
          type: string
        required: true
      tags:
      - products
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedReviewListList'
          description: ''
    post:
      operationId: products_reviews_create
      description: |-
        ``/api/v1/products/{slug}/reviews/`` family.

        GenericViewSet (not ModelViewSet) because the resource is nested
        under product -- we don't want a top-level list view here.
      parameters:
      - in: path
        name: slug
        schema:
          type: string
        required: true
      tags:
      - products
      security:
      - jwtAuth: []
      - {}
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewList'
          description: ''
...
  /api/v1/reviews/{id}/:
    get:
      operationId: reviews_retrieve
      description: '``/api/v1/reviews/{id}/`` family.'
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - reviews
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewList'
          description: ''
    put:
      operationId: reviews_update
      description: '``/api/v1/reviews/{id}/`` family.'
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - reviews
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewList'
          description: ''
    patch:
      operationId: reviews_partial_update
      description: '``/api/v1/reviews/{id}/`` family.'
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - reviews
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewList'
          description: ''
    delete:
      operationId: reviews_destroy
      description: '``/api/v1/reviews/{id}/`` family.'
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - reviews
      security:
      - jwtAuth: []
      responses:
        '204':
          description: No response body
  /api/v1/reviews/{id}/helpful/:
    post:
      operationId: reviews_helpful_create
      description: '``/api/v1/reviews/{id}/`` family.'
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - reviews
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewList'
          description: ''
  /api/v1/reviews/can-review/:
    get:
      operationId: reviews_can_review_retrieve
      description: |-
        Return ``{can_review: bool, reason: str|null}``.

        The product is identified by the ``product`` query parameter
        (slug).  ``IsCustomer`` enforces the customer-only role check;
        non-customers receive 403 before the eligibility logic runs.
      tags:
      - reviews
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewList'
          description: ''
  /api/v1/search/:
    get:
      operationId: search_retrieve
      description: Full-text + faceted product search (Module 11 §11.1).
      tags:
      - search
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          description: No response body
...
  /api/v1/vendor/reviews/:
    get:
      operationId: vendor_reviews_list
      description: '``/api/v1/vendor/reviews/`` family.'
      parameters:
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      tags:
      - vendor
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedReviewListList'
          description: ''
  /api/v1/vendor/reviews/{id}/reply/:
    post:
      operationId: vendor_reviews_reply_create
      description: '``/api/v1/vendor/reviews/`` family.'
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - vendor
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewList'
          description: ''
    delete:
      operationId: vendor_reviews_reply_destroy
      description: '``/api/v1/vendor/reviews/`` family.'
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - vendor
      security:
      - jwtAuth: []
      responses:
        '204':
          description: No response body
...
components:
  schemas:
...
    PaginatedAdminReviewListList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/AdminReviewList'
...
    PaginatedReviewListList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/ReviewList'
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT